chatly-agent/pyproject.toml
[project]
name = "agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
	"fastapi>=0.111.0",
	"uvicorn[standard]>=0.30.0",
	"pydantic-settings>=2.3.0",
	"motor>=3.7.1",
	"pymongo>=4.15.0",
	"python-jose[cryptography]>=3.3.0",
	"passlib[bcrypt]>=1.7.4",
	"bcrypt<4",
	"python-multipart>=0.0.9",
	"langgraph>=0.2.0",
	"langchain>=0.2.0",
	"langchain-groq>=0.1.0",
	"langserve>=0.2.0",
	"pypdf>=6.9.1",
	"python-docx>=1.2.0",
	"langchain-text-splitters>=1.1.1",
	"langchain-huggingface>=1.2.1",
	"huggingface-hub>=1.7.2",
	"minio>=7.2.20",
	"qdrant-client>=1.15.1",
	"numpy>=2.4.3",
	"langchain-community>=0.4.1",
	"tavily-python>=0.7.23",
	"langchain-tavily>=0.2.17",
	"slowapi>=0.1.9",
]

[dependency-groups]
dev = [
	"pytest>=8.0.0",
	"pytest-asyncio>=0.23.0",
	"pytest-cov>=5.0.0",
	"respx>=0.21.0",
	"ruff>=0.6.0",
	"mypy>=1.10.0",
	"httpx>=0.27.0",
	"mongomock-motor>=0.0.36",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short -q"
pythonpath = ["."]

[tool.coverage.run]
source = ["app"]
omit = ["app/main.py", "app/config.py"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
